#8 Bruchmengen: Mengen-Engine von Integer auf Decimal (NUMERIC(14,3))
Alle Bestandsmengen sind ab jetzt Dezimalzahlen – damit funktionieren
Abrechnung/Reservierung/FIFO für JEDE Einheit (kg, m², m³, l) inkl.
Bruchmengen, nicht mehr nur ganze Stück.

Backend
- Spalten instances.quantity/reserved_quantity, orders.quantity,
  order_lines.quantity, purchase_orders.quantity, sales.quantity →
  NUMERIC(14,3) (Migration 055 + idempotenter ALTER-Safety-Net in main.py).
- Neue EINE Mengen-Stelle services/quantity.py (to_qty/qty_sum/is_whole/qty_key):
  kapselt Umwandlung/Rundung (3 NK, exakt, kein float).
- reservation/inventory/subject/process/resource/recovery/scrap/serialization
  rechnen durchgängig mit Decimal. Die Reservierungs-Map (instances.reservations)
  speichert Mengen JSON-sicher als String; events.emit serialisiert Decimal→float.
- Einzelteil-Artikel (serialization='unit') nur GANZE Stück (Router-Check gegen
  die Serialisierung); Chargen (batch) tragen Bruchmengen. required_sample liefert
  ganze Proben (ceil, auch für 2.5 kg).
- API-Rand: Mengen als float (TS number – Frontend unverändert). AI-Tools geben
  Mengen als Zahl (kein int-Truncate mehr).

Frontend
- Menge/Ressourcen-Zeile/Teil-Verschrottung akzeptieren Nachkommastellen
  (kein Math.trunc mehr; step="any"/inputMode="decimal").

Tests: tests/test_fractional_quantities.py (Decimal-Roundtrip, FIFO, Stichprobe).

+++ backend/app/models/instance.py
@@ -1,7 +1,8 @@
 from datetime import datetime
+from decimal import Decimal
 from typing import Optional
 
-from sqlalchemy import BigInteger, DateTime, Integer, String
+from sqlalchemy import BigInteger, DateTime, Numeric, String
 from sqlalchemy.dialects.postgresql import JSONB
 from sqlalchemy.orm import Mapped, mapped_column
 
@@ -29,7 +30,10 @@ class Instance(Base, TimestampMixin):
     order_id: Mapped[int] = mapped_column(BigInteger, index=True, nullable=False)
 
     kind: Mapped[str] = mapped_column(String(10), default="unit", nullable=False)   # unit | batch
-    quantity: Mapped[int] = mapped_column(Integer, default=1, nullable=False)
+    # Menge als Dezimalzahl (NUMERIC(14,3)): ein Einzelteil trägt 1, eine Charge die
+    # Bestellmenge – für kg/m²/m³/l auch als **Bruchmenge** (2.5 kg, 0.75 m²), nicht nur
+    # ganze Stück. Arithmetik ausschliesslich über ``services/quantity.py`` (exakt, kein float).
+    quantity: Mapped[Decimal] = mapped_column(Numeric(14, 3), default=1, nullable=False)
     serial_number: Mapped[Optional[str]] = mapped_column(String(60), nullable=True)
 
     # ZWEI getrennte Achsen statt eines überladenen qc_status:
@@ -56,7 +60,8 @@ class Instance(Base, TimestampMixin):
     # 1000 Schrauben, von der 30 reserviert sind, bleibt also mit 970 frei verfügbar –
     # **ohne** dass eine zweite Instanz mit eigener Nummer entsteht.
     reservations: Mapped[Optional[dict]] = mapped_column(JSONB, nullable=True)
-    reserved_quantity: Mapped[int] = mapped_column(Integer, default=0, server_default="0", nullable=False)
+    reserved_quantity: Mapped[Decimal] = mapped_column(
+        Numeric(14, 3), default=0, server_default="0", nullable=False)
 
     # Einzel-Reservierungs-Zeiger (Altfeld / Schnellprüfung): gesetzt, solange genau EIN
     # Auftrag die Instanz (teil-)reserviert. Massgeblich ist die ``reservations``-Map.
